feat: move google docs function to matching module
