+++ app/integrations/google_workspace/google_docs.py
@@ -12,6 +12,8 @@ Functions:
     get(document_id: str) -> dict:
         Gets a document from Google Docs and returns the document resource.
 """
+import logging
+import re
 from integrations.google_workspace.google_service import (
     get_google_service,
     handle_google_api_errors,
@@ -67,3 +69,20 @@ def get(document_id: str) -> dict:
     service = get_google_service("docs", "v1")
     result = service.documents().get(documentId=document_id).execute()
     return result
+
+
+def extract_google_doc_id(url):
+    # if the url is empty or None, then log an error
+    if not url:
+        logging.error("URL is empty or None")
+        return None
+
+    # Regular expression pattern to match Google Docs ID
+    pattern = r"https://docs.google.com/document/d/([a-zA-Z0-9_-]+)/"
+
+    # Search in the given text for all occurences of pattern
+    match = re.search(pattern, url)
+    if match:
+        return match.group(1)
+    else:
+        return None
