+++ app/tests/integrations/google_workspace/test_google_docs.py
@@ -52,3 +52,31 @@ def test_get_returns_document_resource(get_google_service_mock):
     assert document["documentId"] == "test_document_id"
     assert document["title"] == "test_document"
     assert document["body"] == {"content": [{}]}
+
+
+def test_extract_googe_doc_id_valid_google_docs_url():
+    url = "https://docs.google.com/document/d/1aBcD_efGHI/edit"
+    assert google_docs.extract_google_doc_id(url) == "1aBcD_efGHI"
+
+
+def test_extract_googe_doc_id_oogle_docs_url_with_parameters():
+    url = "https://docs.google.com/document/d/1aBcD_efGHI/edit?usp=sharing"
+    assert google_docs.extract_google_doc_id(url) == "1aBcD_efGHI"
+
+
+def test_extract_googe_doc_id_non_google_docs_url():
+    url = "https://www.example.com/page/d/1aBcD_efGHI/other"
+    assert google_docs.extract_google_doc_id(url) is None
+
+
+def test_extract_googe_doc_id_invalid_url_format():
+    url = "https://docs.google.com/document/1aBcD_efGHI"
+    assert google_docs.extract_google_doc_id(url) is None
+
+
+def test_extract_googe_doc_id_empty_string():
+    assert google_docs.extract_google_doc_id("") is None
+
+
+def test_extract_googe_doc_id_none_input():
+    assert google_docs.extract_google_doc_id(None) is None
